feat(run): new script for development runs
The idea is to use this script for testing some of the
repositories together. This initial implementation allows
one to run the desktop build with the CLI. This specific
invocation of the script is only meant for quick development,
therefore we just build directly using golang. This is one
of the main feature I'd expect from a monorepository.