refactor: introduce the concept of workflow
A workflow is a YAML file that describes how to do something
as a sequence of steps (this may sound quite familiar).

An action is a script called by a workflow.

Actions and workflows live inside ./tools/run.d/.

Each directory inside ./tools/run.d should contain a workflow.yml
file (which is a workflow), an actions.bash file (which is an
action) or both.

You still use the ./tools/run script to run things and the overall
semantic of calling this script has not changed.

The script runs workflows. You cannot directly run actions.

Actions are reusable pieces of code. Workflows are the top-level
abstraction that you can invoke.

We run workflows using the ./tools/run-workflow Go tool.

This tool is a tiny wrapper around ./tools/run-action/run, which
is a bash script.

Previously existing actions map basically 1:1 to an action.bash.

Some actions used to directly call other actions. Now this logic is
inside of the workflow.yml file.

Previous scripts at ./tools/run.d have now become workflows and/or
actions depending on what they were doing.

Why am I doing this? Two reasons:

1. being able to separate the workdir of each workflow gives us much
more confidence that the contents of a build are correct;

2. we're now able to restart a workflow from the step that failed or
just re-run part of a workflow, which is quite useful in case of
failure as well as you just need to re-run a single part of it.