feat: make actions batch by default
This allows them to store logs. In turn, logs are useful to
evaluate what a build did, and when it did it.

Also allow actions to be interactive. For example, the action
running debian:stable in Docker must be interactive because
we're supposed to interact with the Docker's stdin.

While there, log the current status of the monorepo when we
start an action, so we know what we're building more precisely.