fix(t/reset): ensure we're not confused by emojis
It seems using run inside the subshell is not a good idea given
that now we're emitting emojis on the stdout.

Hence, just basically avoid logging commands in subshells.

(This may actually be a good reason to revert back to the previous
behavior but not using pretty-printing in subshelles seems more
robust as a strategy so I'll do just that for now.)