feat(filternet): control iptables and netem
This new command implements a low-level version of jafar that does
not manage any kind of proxy. Compared to jafar, this tool is (1) more
low-level and (2) more flexible. You can basically configure any
iptables rule you wish. Of course, this also means that this command
is less user friendly than jafar is. So, what's the catch?

Basically, the idea is to have a simple low-level tool that allows
us to do the following operations:

1. configure netem policies (which jafar cannot do);

2. set arbitrary iptables rules (including blocking specific
endpoints rather than just IP addresses, which jafar cannot do);

3. organize redirection of selected flows to proxies (which
jafar currently does) without implementing the proxies.

Then, as a subsequent commit, we'll reimplement Jafar as a set
of proxies plus the knowledge of how to invoke this low-level
command for shaping, blocking, and force-proxing.

By having two separate commands, we'll separate the details of
how to interact with iptables (in this command) with how we want
to actually do censorship at a high level.

In turn, this opens up the possibility of choosing to use
enother tool to configure censorship, e.g., `miniooni --censor`,
which isn't of course as good as blocking using iptables but
will work on Windows and macOS, for better QA.

While there, let us also add an action for building filternet
and a workflow that shows some common use cases in action.