fix(libcore.bash): always log on the standard error
So, this experiment on logging on the standard output has created
issues yesterday with the reset command (see 4816db2ff) and now is
breaking the generation of the Android keystore.

I guess two data points are enough to consider this experiment
failed and we should instead always use the standard error.

I'm not going to revert 4816db2ff since the code is working fine
anyway and it doesn't matter much to have that log back.