feat: use canonical directory for Android SDK
Saves quite some bytes on the disk if you happen to already
have an Android Studio setup on the same system.

To make this possible, we need to decouple the location of the
Android SDK from the general SDK dir.

Now `sdk_base_dir` is only valid for Go and we use a different
root directory for Android, which is fine.

Make sure we have a workflow to setup android (is useful).

While there, make sure we're using the latest version of the
tools we install and document how to check for updates.

While there, add my preferred editorconfig.