fix(setup-android): run sdkmanager --update on each run
Won't hurt because it will fetch the latest version of the packages
we have installed (as far as I understand this command :-).

For context, I just run manually `sdkmanager --update` on an old
installation and it upgraded the command line tools.

So, this command seems to nice to have in the build pipeline.