fix(t/r-w): better scope for interactive functionality
We need to run interactively when a process needs the stdin. This
happens for some docker or sudo invocations.

In all other cases, interactive was an hack to see the stdout/stderr.

So, remove the `-i` cmdline flag that served the purpose of seeing
stdout/stderr and instead tee stdout/stderr both to a logfile and to
the current process' stdout and stderr.

However, let workflows declare whether they're interactive.